Proc Macros + More Threading
Add a basic pro macro to handle loading lua scripts. Also ported the scope generation over to a thread. That port isn't 100% up to date when it comes to handling variables/keywords that are imbedded; there was a memory leak where items were being duplicated a bunch, and it's not super trivial to fix.