Refactoring
Refactored the main runtime branches to be threaded manually for improved performance. Also improved scrolling and event polling. A new rendering framework is on the way which should improve the performance more (currently, the draw call is happening very often and I'm not sure if the current system is smart enough to not render when non-updated).