Improved LSP Interfacing
Improved some of the backend elements for the Rust-Analyzer integration. Currently, the LSP is only being initialized and closed (the closing has been verified to ensure no long-term memory leaks from accumulating instances of the binary that never get closed. A lot of care was put into ensuring proper exiting regardless of the reason for the app's closure; an error, or the user, doesn't matter). Further functionality is in the works.